// src/pocketbase-types.ts
export type IsoDateString = string
// ...
export type ProjectsRecord = {
  created?: IsoDateString
  description: string
  id: string
  thumbnail?: string
  title: string
  updated?: IsoDateString
  url?: string
}